Bug fixes, Integer para Long.

+++ src/br/ucsal/Testable.java
@@ -3,8 +3,8 @@ package br.ucsal;
 public interface Testable {
 	
 	public <T extends Comparable<T> > void sort(T[] arr);
-	public Integer getNumComparacao();
-	public Integer getNumMovimento();
+	public Long getNumComparacao();
+	public Long getNumMovimento();
 
 	String getNome();
